+++ sdk/typespec/typespec_client_core/src/lib.rs
@@ -23,3 +23,8 @@ pub mod xml;
 pub use crate::error::{Error, Result};
 pub use bytes::Bytes;
 pub use uuid::Uuid;
+
+#[doc(hidden)]
+mod __private {
+    pub trait Sealed {}
+}